backend/requirements-minimal.txt
Flask==2.3.3
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
python-dotenv==1.0.0
firebase-admin==6.5.0
PyJWT==2.8.0
cryptography==42.0.2
bcrypt==4.3.0
Werkzeug==3.0.1
python-dateutil==2.8.2
requests==2.32.3
urllib3==2.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
qrcode==7.3.1
pyotp==2.9.0

